requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.24
alembic==1.13.1
asyncpg==0.29.0